fix(qemu,cli,dbg): 切回 QEMU 自带 EDK2 firmware + cli list/dbg mouse 解析修复 [EDK2 boot fix + CLI/dbg fixes]
- scripts/qemu-build.sh: fetch_edk2_firmware 不再下载 retrage edk2-nightly
  (与 Ubuntu 24.04 arm64 ISO 不兼容, EDK2 splash 后不试 boot device 卡死),
  改为校验 + 复用 QEMU 自带 pc-bios/edk2-aarch64-code.fd.bz2 (kraxel build,
  跟 brew QEMU 同源); vars 模板从 QEMU 自带 edk2-arm-vars.fd 复制后 padding
  到 64MB. 不再需要外网下载 firmware.
- hvm-cli list: DISK(main) 列对 qcow2 显示 0.0/0Gi (qcow2 stat.st_size 是文件
  实际占用而非 guest 看到的 virtual size). 改读 config.disks[0].sizeGiB.
- hvm-dbg mouse: parsePair 之前返 Double 字符串化成 "640.0", host 端 Int()
  解析失败报"必须是整数". 改成 Int.
- CLAUDE.md / docs/QEMU_INTEGRATION.md: 同步 EDK2 firmware 来源说明.

经 /Volumes/重要盘/ISO/ubuntu-24.04.4-live-server-arm64.iso 端到端验证: EDK2 →
GRUB → kernel → systemd → Subiquity 完整启动, /dev/vda 64.000G 主盘识别正常,
hvm-dbg screenshot/ocr/find-text/key/mouse/wait/console/status/boot-progress
逐项 ok, hvm-cli stop ACPI 优雅关机 + run/ socket 全部清理.

+++ app/Sources/hvm-dbg/Commands/MouseCommand.swift
@@ -88,11 +88,13 @@ struct MouseCommand: AsyncParsableCommand {
         }
     }
 
-    private func parsePair(_ s: String, label: String) throws -> (Double, Double) {
+    private func parsePair(_ s: String, label: String) throws -> (Int, Int) {
+        // host 侧 (QemuHostState.parseXY / DbgOps.handleMouse) 只接受整数像素 — 用
+        // Double 解析会把 "640.0" 发过去, host 端 Int(xs) 返回 nil 报"必须是整数".
         let parts = s.split(separator: ",")
         guard parts.count == 2,
-              let x = Double(parts[0].trimmingCharacters(in: .whitespaces)),
-              let y = Double(parts[1].trimmingCharacters(in: .whitespaces)) else {
+              let x = Int(parts[0].trimmingCharacters(in: .whitespaces)),
+              let y = Int(parts[1].trimmingCharacters(in: .whitespaces)) else {
             throw HVMError.config(.invalidEnum(field: label, raw: s, allowed: ["x,y"]))
         }
         return (x, y)
